feat(DEV-12955): remove @team-monite/sdk-themes
