--- packages/sdk-themes/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@team-monite/sdk-themes",
-  "version": "0.3.0",
-  "private": true,
-  "scripts": {
-    "dev": "yarn clean && tsc --watch",
-    "build": "yarn clean && tsc",
-    "clean": "rimraf dist"
-  },
-  "type": "module",
-  "exports": {
-    "./package.json": "./package.json",
-    ".": "./dist/index.js",
-    "./*": "./dist/*.js"
-  },
-  "typesVersions": {
-    "*": {
-      "*": [
-        "dist/*"
-      ]
-    }
-  },
-  "dependencies": {
-    "@mui/material": "~5.15.14"
-  },
-  "devDependencies": {
-    "rimraf": "~5.0.5",
-    "typescript": "~5.5.4"
-  },
-  "files": [
-    "dist",
-    "src"
-  ]
-}
