--- packages/sdk-themes/src/index.ts
@@ -1,10 +1,0 @@
-export {
-  moniteLight,
-  moniteDark,
-  paletteDark,
-  paletteLight,
-  defaultMoniteComponents,
-  defaultMoniteTypography,
-} from './themes/monite.js';
-
-export { materialLight, materialDark } from './themes/material.js';
