--- packages/sdk-themes/src/themes/material.ts
@@ -1,5 +1,0 @@
-import type { ThemeOptions } from '@mui/material/styles/createTheme.js';
-
-export const materialLight: ThemeOptions = { palette: { mode: 'light' } };
-
-export const materialDark: ThemeOptions = { palette: { mode: 'dark' } };
