--- packages/sdk-themes/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2016",
-    "module": "NodeNext",
-    "outDir": "dist",
-    "baseUrl": ".",
-    "strict": true,
-    "noEmitOnError": true,
-    "moduleResolution": "NodeNext",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "declaration": true,
-    "declarationMap": true,
-    "skipLibCheck": true,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"]
-}
